Move NotifyClusterConnectionLost from (static) GrainClient to OutsideRuntimeClient
