+++ src/Orleans/Orleans.csproj
@@ -82,6 +82,7 @@
     <Compile Include="CodeGeneration\GenericMethodInvoker.cs" />
     <Compile Include="Configuration\GrainServiceConfiguration.cs" />
     <Compile Include="Core\GrainCasterFactory.cs" />
+    <Compile Include="Core\IClusterConnectionStatusListener.cs" />
     <Compile Include="Core\IInternalGrainFactory.cs" />
     <Compile Include="Core\IInternalClusterClient.cs" />
     <Compile Include="Core\IClusterClient.cs" />
